+++ android/vendored/unversioned/react-native-reanimated/android/src/main/java/com/swmansion/reanimated/nodes/FinalNode.java
@@ -1,0 +1,5 @@
+package com.swmansion.reanimated.nodes;
+
+public interface FinalNode {
+  void update();
+}